cnss-core: resolve compile warning
1. force convert between u64 and dma_addr_t
2. use "zx" or "zu" for size_t in printk

+++ mhi/mhi_bhi.c
@@ -53,7 +53,7 @@ static int bhi_alloc_bhie_xfer(struct mhi_device_ctxt *mhi_dev_ctxt,
 	struct bhie_mem_info *bhie_mem_info, *info = NULL;
 
 	mhi_log(mhi_dev_ctxt, MHI_MSG_INFO,
-		"Total size:%lu total_seg:%d seg_size:%lu\n",
+		"Total size:%zu total_seg:%d seg_size:%zu\n",
 		size, segments, seg_size);
 
 	sg_list = kcalloc(segments, sizeof(*sg_list), GFP_KERNEL);
@@ -85,7 +85,7 @@ static int bhi_alloc_bhie_xfer(struct mhi_device_ctxt *mhi_dev_ctxt,
 			(info->phys_addr - info->dma_handle);
 		mhi_log(mhi_dev_ctxt, MHI_MSG_INFO,
 			"Seg:%d unaligned Img: 0x%llx aligned:0x%llx\n",
-			i, info->dma_handle, info->phys_addr);
+			i, (u64)info->dma_handle, (u64)info->phys_addr);
 	}
 
 	sg_init_table(sg_list, segments);
@@ -140,9 +140,9 @@ static int bhi_alloc_pbl_xfer(struct mhi_device_ctxt *mhi_dev_ctxt,
 	mem_info->aligned = mem_info->pre_aligned + (mem_info->phys_addr -
 						     mem_info->dma_handle);
 	mhi_log(mhi_dev_ctxt, MHI_MSG_INFO,
-		"alloc_size:%lu image_size:%lu unal_addr:0x%llx0x al_addr:0x%llx\n",
+		"alloc_size:%zu image_size:%zu unal_addr:0x%llx0x al_addr:0x%llx\n",
 		mem_info->alloc_size, mem_info->size,
-		mem_info->dma_handle, mem_info->phys_addr);
+		(u64)mem_info->dma_handle, (u64)mem_info->phys_addr);
 
 	return 0;
 }
@@ -666,7 +666,7 @@ int bhi_probe(struct mhi_device_ctxt *mhi_dev_ctxt)
 		fw_info->segment_size <<= 1;
 
 	mhi_log(mhi_dev_ctxt, MHI_MSG_INFO,
-		"max sbl image size:%lu segment size:%lu\n",
+		"max sbl image size:%zu segment size:%zu\n",
 		fw_info->max_sbl_len, fw_info->segment_size);
 
 #ifdef CONFIG_NAPIER_X86
